requirements.txt
openai>=1.40.0
pydantic>=2.7.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
rich>=13.7.1
